requirements.txt
numpy==1.26.4
pandas==2.2.3
pyarrow==17.0.0
scipy==1.13.1
scikit-learn==1.5.2
cvxpy==1.5.3
yfinance==0.2.40
pandas-datareader==0.10.0
matplotlib==3.9.2
quantstats==0.0.62
pyyaml==6.0.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
tqdm==4.66.4
pytest==8.3.2
